[TD] Cast both Random_Pick args, fixes 64-bit

+++ tiberiandawn/cell.cpp
@@ -2086,7 +2086,7 @@ bool CellClass::Goodie_Check(FootClass* object, bool check_steel)
                 }
 
                 while (what == -1) {
-                    what = _what[Random_Pick((unsigned)0, sizeof(_what) / sizeof(_what[0]) - 1)];
+                    what = _what[Random_Pick((unsigned)0, (unsigned)(sizeof(_what) / sizeof(_what[0]) - 1))];
 
                     if (what == REVEAL && object->House->IsVisionary)
                         what = -1;
@@ -2258,7 +2258,8 @@ bool CellClass::Goodie_Check(FootClass* object, bool check_steel)
                                                   INFANTRY_E5,
                                                   INFANTRY_E7,
                                                   INFANTRY_RAMBO};
-                    InfantryTypeClass::As_Reference(_inf[Random_Pick((unsigned)0, sizeof(_inf) / sizeof(_inf[0]) - 1)])
+                    InfantryTypeClass::As_Reference(
+                        _inf[Random_Pick((unsigned)0, (unsigned)(sizeof(_inf) / sizeof(_inf[0]) - 1))])
                         .Create_And_Place(Cell_Number(), object->Owner());
                 }
                 return (false);
